+++ models.py
@@ -973,7 +973,7 @@ class TopDownRNNG(nn.Module):
       reduce_context = self.stack_top_h(reduce_states)
       reduce_context = self.stack_to_hidden(reduce_context)
       new_child, _, _ = self.composition(children, ch_lengths, nt, nt_id, reduce_context)
-      new_stack_input[reduces] = new_child
+      new_stack_input[reduces] = new_child.float()
 
     new_stack_input[shift_idx] = shifted_embs
